Modify server code to continuously render static HTML
