+++ lib/render.js
@@ -7,16 +7,16 @@ var data = require('./data');
 
 var TOTAL_TRAINS = 2986;
 var JADE_TEMPLATE_PATH = path.join(process.cwd(), 'views/index.jade');
-var OUTPUT_HTML_PATH = path.join(process.cwd(), 'public/output.html');
+var OUTPUT_HTML_PATH = path.join(process.cwd(), 'public/index.html');
 
 module.exports = function () {
   data.getServiceAlerts(function (err, serviceAlerts) {
     data.getTripUpdates(function (err, tripUpdates) {
       fs.readFile(JADE_TEMPLATE_PATH, 'utf8', function (err, data) {
-
         if (err) throw err;
-        var fn = jade.compile(data);
-        var html = fn({
+
+        var template = jade.compile(data);
+        var html = template({
           serviceAlerts: serviceAlerts.entity,
           tripUpdates: tripUpdates.entity.map(function (tripUpdate) {
             var totalDelayInSeconds = tripUpdate.trip_update.stop_time_update.reduce(function (total, stopTimeUpdate) {
@@ -32,9 +32,7 @@ module.exports = function () {
           percentageOfDelays: (tripUpdates.entity.length / TOTAL_TRAINS * 100).toFixed(2)
         });
 
-        fs.writeFile(OUTPUT_HTML_PATH, html, function (err, data) {
-          console.log('yeah baby done');
-        });
+        fs.writeFile(OUTPUT_HTML_PATH, html);
       });
     });
   });
